F OpenNebula/one#5502: Improve container clean (#1418)
(cherry picked from commit 88d720886537ba2232eed64cd5e771088ab321fa)

+++ src/vmm_mad/remotes/lib/lxc/container.rb
@@ -129,10 +129,10 @@ class Container
         end
 
         # Clean bindpoint
-        FileUtils.rm_rf(@one.bind_folder)
+        FileUtils.rm_rf(@one.bind_folder) if Dir.exist?(@one.bind_folder)
 
         # Destroy container
-        @client.destroy(@one.vm_name)
+        @client.destroy(@one.vm_name) if @client.list.include?(@one.vm_name)
     end
 
     #---------------------------------------------------------------------------
